+++ remotestate-py/src/remotestate/__init__.py
@@ -2,7 +2,7 @@
 
 from importlib.metadata import version
 
-from .path import Index, Path, PathSegment, Property
+from . import path
 from .service import Service, action, query
 from .serve import serve
 from .store import Store
@@ -15,8 +15,5 @@ __all__ = [
     "action",
     "query",
     "serve",
-    "Index",
-    "Path",
-    "PathSegment",
-    "Property",
+    "path",
 ]